rename factory to def, move some unit test files
