+++ stream/src/main/java/org/apache/kafka/streaming/kstream/WindowDef.java
@@ -21,5 +21,5 @@ public interface WindowDef<K, V> {
 
     String name();
 
-    Window<K, V> build();
+    Window<K, V> define();
 }
