+++ stream/src/main/java/org/apache/kafka/streaming/kstream/internals/KStreamFilter.java
@@ -19,9 +19,9 @@ package org.apache.kafka.streaming.kstream.internals;
 
 import org.apache.kafka.streaming.processor.Processor;
 import org.apache.kafka.streaming.kstream.Predicate;
-import org.apache.kafka.streaming.processor.ProcessorFactory;
+import org.apache.kafka.streaming.processor.ProcessorDef;
 
-class KStreamFilter<K, V> implements ProcessorFactory {
+class KStreamFilter<K, V> implements ProcessorDef {
 
     private final Predicate<K, V> predicate;
     private final boolean filterOut;
@@ -32,7 +32,7 @@ class KStreamFilter<K, V> implements ProcessorFactory {
     }
 
     @Override
-    public Processor build() {
+    public Processor define() {
         return new KStreamFilterProcessor();
     }
