+++ stream/src/main/java/org/apache/kafka/streaming/kstream/internals/KStreamFlatMap.java
@@ -20,9 +20,9 @@ package org.apache.kafka.streaming.kstream.internals;
 import org.apache.kafka.streaming.kstream.KeyValue;
 import org.apache.kafka.streaming.kstream.KeyValueFlatMap;
 import org.apache.kafka.streaming.processor.Processor;
-import org.apache.kafka.streaming.processor.ProcessorFactory;
+import org.apache.kafka.streaming.processor.ProcessorDef;
 
-class KStreamFlatMap<K1, V1, K2, V2> implements ProcessorFactory {
+class KStreamFlatMap<K1, V1, K2, V2> implements ProcessorDef {
 
     private final KeyValueFlatMap<K1, V1, K2, V2> mapper;
 
@@ -31,7 +31,7 @@ class KStreamFlatMap<K1, V1, K2, V2> implements ProcessorFactory {
     }
 
     @Override
-    public Processor build() {
+    public Processor define() {
         return new KStreamFlatMapProcessor();
     }
