+++ stream/src/main/java/org/apache/kafka/streaming/kstream/internals/KStreamMapValues.java
@@ -19,9 +19,9 @@ package org.apache.kafka.streaming.kstream.internals;
 
 import org.apache.kafka.streaming.processor.Processor;
 import org.apache.kafka.streaming.kstream.ValueMapper;
-import org.apache.kafka.streaming.processor.ProcessorFactory;
+import org.apache.kafka.streaming.processor.ProcessorDef;
 
-class KStreamMapValues<K1, V1, V2> implements ProcessorFactory {
+class KStreamMapValues<K1, V1, V2> implements ProcessorDef {
 
     private final ValueMapper<V1, V2> mapper;
 
@@ -30,7 +30,7 @@ class KStreamMapValues<K1, V1, V2> implements ProcessorFactory {
     }
 
     @Override
-    public Processor build() {
+    public Processor define() {
         return new KStreamMapProcessor();
     }
