+++ stream/src/main/java/org/apache/kafka/streaming/processor/ProcessorDef.java
@@ -17,7 +17,7 @@
 
 package org.apache.kafka.streaming.processor;
 
-public interface ProcessorFactory {
+public interface ProcessorDef {
 
-    Processor build();
+    Processor define();
 }
