+++ stream/src/test/java/org/apache/kafka/streaming/kstream/internals/KStreamFlatMapTest.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.kafka.streaming.internals;
+package org.apache.kafka.streaming.kstream.internals;
 
 import org.apache.kafka.common.serialization.IntegerDeserializer;
 import org.apache.kafka.common.serialization.StringDeserializer;
@@ -30,7 +30,9 @@ import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
 
-public class KStreamMapTest {
+import java.util.ArrayList;
+
+public class KStreamFlatMapTest {
 
     private String topicName = "topic";
 
@@ -39,35 +41,40 @@ public class KStreamMapTest {
     private StringDeserializer valDeserializer = new StringDeserializer();
 
     @Test
-    public void testMap() {
+    public void testFlatMap() {
 
-        KeyValueMapper<Integer, String, String, Integer> mapper =
-            new KeyValueMapper<Integer, String, String, Integer>() {
+        KeyValueMapper<Integer, String, String, Iterable<String>> mapper =
+            new KeyValueMapper<Integer, String, String, Iterable<String>>() {
                 @Override
-                public KeyValue<String, Integer> apply(Integer key, String value) {
-                    return KeyValue.pair(value, key);
+                public KeyValue<String, Iterable<String>> apply(Integer key, String value) {
+                    ArrayList<String> result = new ArrayList<String>();
+                    for (int i = 0; i < key; i++) {
+                        result.add(value);
+                    }
+                    return KeyValue.pair(Integer.toString(key * 10), (Iterable<String>) result);
                 }
             };
 
         final int[] expectedKeys = new int[]{0, 1, 2, 3};
 
         KStream<Integer, String> stream;
-        MockProcessor<String, Integer> processor;
+        MockProcessor<String, String> processor;
 
         processor = new MockProcessor<>();
         stream = topology.<Integer, String>from(keyDeserializer, valDeserializer, topicName);
-        stream.map(mapper).process(processor);
+        stream.flatMap(mapper).process(processor);
 
         for (int i = 0; i < expectedKeys.length; i++) {
             ((KStreamSource<Integer, String>) stream).source().process(expectedKeys[i], "V" + expectedKeys[i]);
         }
 
-        assertEquals(4, processor.processed.size());
+        assertEquals(6, processor.processed.size());
 
-        String[] expected = new String[]{"V0:0", "V1:1", "V2:2", "V3:3"};
+        String[] expected = new String[]{"10:V1", "20:V2", "20:V2", "30:V3", "30:V3", "30:V3"};
 
         for (int i = 0; i < expected.length; i++) {
             assertEquals(expected[i], processor.processed.get(i));
         }
     }
+
 }
