+++ stream/src/test/java/org/apache/kafka/streaming/processor/internals/StreamGroupTest.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.kafka.streaming.internals;
+package org.apache.kafka.streaming.processor.internals;
 
 import org.apache.kafka.clients.consumer.ConsumerRecord;
 import org.apache.kafka.streaming.processor.TimestampExtractor;
